more refactoring of files
