+++ apps/frontend/src/components/questions/AddQuestionModal.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { Question } from "./Question";
-import { Complexity } from "./enums/Complexity";
-import { Category } from "./enums/Category";
+import { Question } from "../Question";
+import { Complexity } from "../enums/Complexity";
+import { Category } from "../enums/Category";
 import styles from "/styles/modal.module.css";
 
 type AddQuestionModalProps = {
